tests/unit/requirements.txt
mock
pyyaml
coverage
six
urllib3
